scriptFix: remove = unused variable, else statement, Add=check dir exist, counter file deleted, successful message into the terminal

+++ DailyLogs/November-2025/23-nov/script2/script2.py
@@ -9,15 +9,26 @@ if len(sys.argv) < 2:
     print(f"Usage: python3 script.py <arguments>")
     sys.exit(2) # invalid arguments
 
-dir_name = sys.argv[1] #dir name
-dir_path = os.path.join(dir_name) # dir name
-logging.info("Cleaning process begun...")
+dir_path = sys.argv[1] #dir name
 
+# check if the directory exist
+if not os.path.isfile(dir_path):
+    logging.error("Directory doesn't exist.",dir_path)
+    print("Directory doesn't exist.")
+    sys.exit(1) # missing file
+
+logging.info("Cleaning process begun...")
+deleted_files = 0
+
+# loging to check for .tmp file recursively
 for root, dirs, files in os.walk(dir_path):
     for file in files:
         if file.endswith(".tmp"):
             fullpath = os.path.join(root,file)
             os.remove(fullpath)
+            deleted_files += 1
             logging.info(f"{fullpath} been removed successfully.")
-        else:
-            logging.error(f"Not possible to remvoe.")
+       
+logging.info(f"Successfully cleaned {deleted_files} files.")
+print(f"Successfully cleaned {deleted_files} files.")
+
